+++ codegen/templates/ethereum_chain.go
@@ -43,14 +43,14 @@ var EthereumChainsByID = map[string]*EthereumChain{
 		DefaultContractName:    "GETH Token",
 		FirehoseEndpoint:       "goerli.eth.streamingfast.io:443",
 	},
-	"Sepolia": {
-		DisplayName:            "Sepolia Testnet",
-		ExplorerLink:           "https://sepolia.etherscan.io",
-		ApiEndpoint:            "https://api-sepolia.etherscan.io",
-		DefaultContractAddress: "0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984",
-		DefaultContractName:    "UNI Token",
-		FirehoseEndpoint:       "sepolia.streamingfast.io:443",
-	},
+	// "Sepolia": {
+	// 	DisplayName:            "Sepolia Testnet",
+	// 	ExplorerLink:           "https://sepolia.etherscan.io",
+	// 	ApiEndpoint:            "https://api-sepolia.etherscan.io",
+	// 	DefaultContractAddress: "0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984",
+	// 	DefaultContractName:    "UNI Token",
+	// 	FirehoseEndpoint:       "sepolia.streamingfast.io:443",
+	// },
 }
 
 func init() {